phase-15-ai-rag/requirements.txt
sentence-transformers
supabase
psycopg2
python-dotenv
